+++ packages/palette/src/elements/Button/index.tsx
@@ -1,2 +1,1 @@
 export * from "./Button"
-export * from "./Button.shared"
