+++ packages/palette/src/elements/RadioGroup/RadioGroup.story.tsx
@@ -1,7 +1,7 @@
 import { action } from "@storybook/addon-actions"
 import { storiesOf } from "@storybook/react"
 import React, { useState } from "react"
-import { BorderBox } from "../BorderBox/BorderBox.ios"
+import { BorderBox } from "../BorderBox"
 import { Radio } from "../Radio/Radio"
 import { RadioGroup } from "./RadioGroup"
 
@@ -21,8 +21,16 @@ storiesOf("Components/RadioGroup", module)
             Toggle default value: {defaultValue}
           </BorderBox>
           <RadioGroup defaultValue={defaultValue}>
-            <Radio value="SHIP" label="Provide shipping address" onSelect={action('onSelect')} />
-            <Radio value="PICKUP" label="Arrange for pickup" onSelect={action('onSelect')} />
+            <Radio
+              value="SHIP"
+              label="Provide shipping address"
+              onSelect={action("onSelect")}
+            />
+            <Radio
+              value="PICKUP"
+              label="Arrange for pickup"
+              onSelect={action("onSelect")}
+            />
           </RadioGroup>
         </>
       )
